+++ src/ethereum/ethereumInfos.ts
@@ -5,6 +5,7 @@ import { avalanche } from './info/avalancheInfo'
 import { base } from './info/baseInfo'
 import { binancesmartchain } from './info/binancesmartchainInfo'
 import { bobevm } from './info/bobevmInfo'
+import { botanix } from './info/botanixInfo'
 import { celo } from './info/celoInfo'
 import { ethDev } from './info/ethDevInfo'
 import { ethereumclassic } from './info/ethereumclassicInfo'
@@ -30,6 +31,7 @@ export const ethereumPlugins = {
   base,
   binancesmartchain,
   bobevm,
+  botanix,
   celo,
   ethDev,
   ethereum,